@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+##
+# Module used to wrap communication with Amazon Cognito
 module Cognito
   ##
   # Class responsible for getting actual user data from Cognito.
@@ -51,7 +53,6 @@ module Cognito
       unless defined? @user_data
         log_action('Getting user')
         @user_data = client.get_user(access_token: access_token)
-        log_successful_call
       end
 
       @user_data
