@@ -1,12 +1,11 @@
 # frozen_string_literal: true
 
-# Be sure to restart your server when you modify this file.
-
 # Configure sensitive parameters which will be filtered from the log file.
 Rails.application.config.filter_parameters += %i[
-  username
-  password
-  confirmation_code
-  vrn
   authenticity_token
+  confirmation_code
+  groups
+  password
+  username
+  vrn
 ]
