@@ -35,15 +35,22 @@ Rails.application.routes.draw do
     end
   end
 
-  get :accessibility_statement, to: 'static_pages#accessibility_statement'
-  get :cookies, to: 'static_pages#cookies'
-  get :privacy_notice, to: 'static_pages#privacy_notice'
+  scope controller: 'static_pages' do
+    get :accessibility_statement
+    get :cookies
+    get :privacy_notice
+  end
 
-  get :health, to: 'application#health'
-  get :build_id, to: 'application#build_id'
+  scope controller: 'application' do
+    get :build_id
+    get :health
+  end
+
+  scope controller: 'errors' do
+    get :service_unavailable
+  end
 
   match '/404', to: 'errors#not_found', via: :all
-  # There is no 422 error page in design systems
   match '/422', to: 'errors#internal_server_error', via: :all
   match '/500', to: 'errors#internal_server_error', via: :all
   match '/503', to: 'errors#service_unavailable', via: :all
