@@ -1,17 +1,15 @@
 # frozen_string_literal: true
 
-# Scenario: View upload page without cookie
 When('I have no authentication cookie') do
   cookie = get_me_the_cookie('_csv_uploader_session')
   expect(cookie).to be_nil
 end
 
-# Scenario: View upload page with cookie that has not expired
 When('I navigate to a Upload page') do
   visit authenticated_root_path
 end
 
-Then('I should enter valid credentials and press the Continue') do
+When('I enter valid credentials') do
   sign_in_user
 end
 
@@ -20,7 +18,6 @@ And('Cookie is created for my session') do
   expect(cookie).to_not be_nil
 end
 
-# Scenario: View upload page with cookie that has not expired
 When('I have authentication cookie that has not expired') do
   visit new_user_session_path
   sign_in_user
@@ -38,7 +35,6 @@ Then('I am redirected to the unauthenticated root page') do
   expect(page).to have_current_path('/')
 end
 
-# Scenario: Sign in with invalid credentials
 When('I enter invalid credentials') do
   fill_in('user_username', with: 'user@example.com')
   fill_in('user_password', with: 'invalid-password')
@@ -51,7 +47,6 @@ Then('I remain on the current page') do
   expect(page).to have_current_path(new_user_session_path)
 end
 
-# Scenario: View upload page with cookie that has expired
 Given('I have authentication cookie that has expired') do
   # set default session_timeout
   Rails.configuration.x.session_timeout = 15
@@ -65,16 +60,18 @@ Given('I have authentication cookie that has expired') do
   expect(cookie[:expires] < Time.current).to be true
 end
 
-# Scenario: Sign out
 Given('I am signed in') do
   sign_in_user
+end
+
+Given('I am signed and not belongs to any group') do
+  sign_in_user(groups: [])
 end
 
 When('I request to sign out') do
   click_link 'Sign out'
 end
 
-# Scenario: Sign in with invalid email format
 When('I enter invalid email format') do
   fill_in('user_username', with: 'user.example.com')
   fill_in('user_password', with: '12345678')
