@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe UploadMailer, type: :mailer do
+describe UploadMailer do
   let(:user) { create_user(email: email) }
   let(:email) { 'test@example.com' }
 
