@@ -2,8 +2,8 @@
 
 require 'rails_helper'
 
-describe 'PasswordsController - POST #change', type: :request do
-  subject(:http_request) { post change_passwords_path, params: params }
+describe 'PasswordsController - POST #change' do
+  subject { post change_passwords_path, params: params }
 
   let(:params) do
     {
@@ -41,22 +41,22 @@ describe 'PasswordsController - POST #change', type: :request do
     end
 
     it 'returns redirect to success page' do
-      http_request
+      subject
       expect(response).to redirect_to(success_passwords_path)
     end
 
     it 'clears password_reset_token' do
-      http_request
+      subject
       expect(session[:password_reset_token]).to be_nil
     end
 
     it 'clears password_reset_username' do
-      http_request
+      subject
       expect(session[:password_reset_username]).to be_nil
     end
 
     it 'updates user lockout data' do
-      http_request
+      subject
       expect(Cognito::Lockout::UpdateUser).to have_received(:call)
         .with(username: username, failed_logins: 0)
     end
@@ -69,12 +69,12 @@ describe 'PasswordsController - POST #change', type: :request do
       end
 
       it 'returns redirect to confirm_reset_passwords_path' do
-        http_request
+        subject
         expect(response).to redirect_to(confirm_reset_passwords_path)
       end
 
       it 'does not update user lockout data' do
-        http_request
+        subject
         expect(Cognito::Lockout::UpdateUser).not_to have_received(:call)
       end
     end
@@ -87,12 +87,12 @@ describe 'PasswordsController - POST #change', type: :request do
       end
 
       it 'returns redirect to confirm_reset_passwords_path' do
-        http_request
+        subject
         expect(response).to redirect_to(confirm_reset_passwords_path)
       end
 
       it 'does not update user lockout data' do
-        http_request
+        subject
         expect(Cognito::Lockout::UpdateUser).not_to have_received(:call)
       end
     end
@@ -100,7 +100,7 @@ describe 'PasswordsController - POST #change', type: :request do
 
   context 'without password_reset_token set' do
     it 'returns redirect to success page' do
-      http_request
+      subject
       expect(response).to redirect_to(success_passwords_path)
     end
   end
