@@ -2,8 +2,8 @@
 
 require 'rails_helper'
 
-describe 'PasswordsController - POST #create', type: :request do
-  subject(:http_request) { post passwords_path, params: params }
+describe 'PasswordsController - POST #create' do
+  subject { post passwords_path, params: params }
 
   let(:params) { {} }
 
@@ -13,7 +13,7 @@ describe 'PasswordsController - POST #create', type: :request do
     let(:user) { create_user }
 
     it 'returns a redirect to root_path' do
-      http_request
+      subject
       expect(response).to redirect_to(root_path)
     end
   end
@@ -22,7 +22,7 @@ describe 'PasswordsController - POST #create', type: :request do
     let(:user) { create_user(aws_status: 'FORCE_NEW_PASSWORD', aws_session: nil) }
 
     it 'returns a redirect to new_user_session_path' do
-      http_request
+      subject
       expect(response).to redirect_to(new_user_session_path)
     end
   end
@@ -37,7 +37,7 @@ describe 'PasswordsController - POST #create', type: :request do
         let(:params) { { user: { password: nil, password_confirmation: 'test' } } }
 
         it 'renders passwords/new template' do
-          expect(http_request).to render_template(:new)
+          expect(subject).to render_template(:new)
         end
       end
 
@@ -45,7 +45,7 @@ describe 'PasswordsController - POST #create', type: :request do
         let(:params) { { user: { password: 'test', password_confirmation: nil } } }
 
         it 'renders passwords/new template' do
-          expect(http_request).to render_template(:new)
+          expect(subject).to render_template(:new)
         end
       end
 
@@ -53,7 +53,7 @@ describe 'PasswordsController - POST #create', type: :request do
         let(:params) { { user: { password: 'test', password_confirmation: 'test1' } } }
 
         it 'renders passwords/new template' do
-          expect(http_request).to render_template(:new)
+          expect(subject).to render_template(:new)
         end
       end
     end
@@ -64,7 +64,7 @@ describe 'PasswordsController - POST #create', type: :request do
       context 'when call to AWS is successful' do
         before do
           allow(Cognito::RespondToAuthChallenge).to receive(:call).and_return(user)
-          http_request
+          subject
         end
 
         it 'returns a redirect to success_passwords_path' do
@@ -77,11 +77,11 @@ describe 'PasswordsController - POST #create', type: :request do
           allow(Cognito::RespondToAuthChallenge).to receive(:call).and_raise(
             NewPasswordException.new({})
           )
-          http_request
+          subject
         end
 
         it 'renders passwords/new template' do
-          expect(http_request).to render_template(:new)
+          expect(subject).to render_template(:new)
         end
 
         it 'does not log out user' do
@@ -94,7 +94,7 @@ describe 'PasswordsController - POST #create', type: :request do
           allow(Cognito::RespondToAuthChallenge).to receive(:call).and_raise(
             Cognito::CallException.new(I18n.t('expired_session'), new_user_session_path)
           )
-          http_request
+          subject
         end
 
         it 'returns a redirect to new_password_path' do
