@@ -2,12 +2,12 @@
 
 require 'rails_helper'
 
-describe 'PasswordsController - GET #new', type: :request do
-  subject(:http_request) { get new_password_path }
+describe 'PasswordsController - GET #new' do
+  subject { get new_password_path }
 
   before do
     sign_in user
-    http_request
+    subject
   end
 
   context 'when user aws_status is OK' do
@@ -31,8 +31,8 @@ describe 'PasswordsController - GET #new', type: :request do
       create_user(aws_status: 'FORCE_NEW_PASSWORD', aws_session: SecureRandom.uuid)
     end
 
-    it 'returns 200' do
-      expect(response).to be_successful
+    it 'returns a 200 OK status' do
+      expect(response).to have_http_status(:ok)
     end
   end
 end
