@@ -2,18 +2,18 @@
 
 require 'rails_helper'
 
-describe 'PasswordsController - GET #reset', type: :request do
+describe 'PasswordsController - GET #reset' do
   let(:user) { User.new }
 
-  subject(:http_request) { get reset_passwords_path }
+  subject { get reset_passwords_path }
 
-  it 'returns 200' do
-    http_request
-    expect(response).to be_successful
+  it 'returns a 200 OK status' do
+    subject
+    expect(response).to have_http_status(:ok)
   end
 
   it 'sets password_reset_token' do
-    http_request
+    subject
     expect(session[:password_reset_token]).not_to be_nil
   end
 end
