@@ -2,13 +2,13 @@
 
 require 'rails_helper'
 
-describe 'UploadController - GET #data_rules', type: :request do
-  subject(:http_request) { get data_rules_upload_index_path }
+describe 'UploadController - GET #data_rules' do
+  subject { get data_rules_upload_index_path }
 
   before { sign_in create_user }
 
   it 'returns a success response' do
-    http_request
+    subject
     expect(response).to have_http_status(:success)
   end
 end
