@@ -2,14 +2,14 @@
 
 require 'rails_helper'
 
-describe 'UploadController - #force_new_password', type: :request do
-  subject(:http_request) { get root_path }
+describe 'UploadController - #force_new_password' do
+  subject { get root_path }
 
   let(:user) { create_user(aws_status: 'FORCE_NEW_PASSWORD') }
 
   before do
     sign_in user
-    http_request
+    subject
   end
 
   context 'when user aws_status is FORCE_NEW_PASSWORD' do
