@@ -2,20 +2,24 @@
 
 require 'rails_helper'
 
-RSpec.describe 'VehiclesController - GET #not_found', type: :request do
-  subject(:http_request) { get not_found_vehicles_path }
+describe 'VehiclesController - GET #not_found' do
+  subject { get not_found_vehicles_path }
 
-  before do
-    sign_in create_user
-    add_to_session(vrn: 'CU57ABC')
-    http_request
+  context 'user belongs to proper group' do
+    before do
+      sign_in create_user
+      add_to_session(vrn: 'CU57ABC')
+      subject
+    end
+
+    it 'returns ok status' do
+      expect(response).to have_http_status(:ok)
+    end
+
+    it 'renders not found page' do
+      expect(response).to render_template(:not_found)
+    end
   end
 
-  it 'returns ok status' do
-    expect(response).to be_successful
-  end
-
-  it 'renders not found page' do
-    expect(response).to render_template(:not_found)
-  end
+  it_behaves_like 'user does not belongs to any group'
 end
