@@ -2,7 +2,7 @@
 
 RSpec.shared_examples 'a static page' do
   it 'returns a success response' do
-    http_request
+    subject
     expect(response).to have_http_status(:success)
   end
 end
